+++ CHANGES.md
@@ -56,6 +56,24 @@ with COGReader(
 * compare dataset bounds and tile bounds in TMS crs in `rio_tiler.io.base.SpatialMixin.tile_exists` method to allow dataset and TMS not compatible with WGS84 crs (https://github.com/cogeotiff/rio-tiler/pull/429)
 * use `httpx` package instead of requests (author @rodrigoalmeida94, https://github.com/cogeotiff/rio-tiler/pull/431)
 * allow **half pixel** `tile_buffer` around the tile (e.g 0.5 -> 257x257, 1.5 -> 259x259) (author @bstadlbauer, https://github.com/cogeotiff/rio-tiler/pull/405)
+* add support for **intervals** colormap (https://github.com/cogeotiff/rio-tiler/pull/439))
+
+```python
+from rio_tiler.colormap import apply_cmap, apply_intervals_cmap
+
+data = numpy.random.randint(0, 255, size=(1, 256, 256))
+cmap = [
+    # ([min, max], [r, g, b, a])
+    ([0, 1], [0, 0, 0, 0]),
+    ([1, 10], [255, 255, 255, 255]),
+    ([10, 100], [255, 0, 0, 255]),
+    ([100, 256], [255, 255, 0, 255]),
+]
+
+data, mask = apply_intervals_cmap(data, cmap)
+# or
+data, mask = apply_cmap(data, cmap)
+```
 
 **breaking changes**